WorldData cleaning
Tweaking a few data structures, Moves the function bodies to the .cpp file and remove and tweak old code. In particular ItemCollection is now a deque and some ints have been replaced with short ints. Comments about a uniqueID being a 512 char string have been removed as we are using the string class now